--- a/deno/lib/__tests__/bigint.test.ts
+++ b/deno/lib/__tests__/bigint.test.ts
@@ -6,11 +6,16 @@ import { util } from "../helpers/util.ts";
 import * as z from "../index.ts";
 
 test("type guard", () => {
-  const stringToNumber = z.string().transform((arg) => BigInt(Number.MAX_SAFE_INTEGER**(arg.length > 1 ? arg.length : 2)));
+  const stringToNumber = z
+    .string()
+    .transform<bigint>((arg) =>
+      BigInt(Number.MAX_SAFE_INTEGER ** (arg.length > 1 ? arg.length : 2))
+    );
 
   const s1 = z.object({
     stringToNumber
   });
+  
   type t1 = z.input<typeof s1>;
 
   const data = { stringToNumber: "asdf" };