--- a/src/__tests__/async-parsing.test.ts
+++ b/src/__tests__/async-parsing.test.ts
@@ -263,11 +263,10 @@ test("enum async parse", async () => {
 enum nativeEnumTest {
   asdf = "qwer",
 }
-// @ts-ignore
-const nativeEnumSchema = z.nativeEnum(nativeEnumTest);
+const nativeEnumSchema = z.nativeEnum<typeof nativeEnumTest>(nativeEnumTest);
 test("nativeEnum async parse", async () => {
   const goodData = nativeEnumTest.asdf;
-  const badData = "asdf";
+  const badData: keyof typeof nativeEnumTest = "asdf";
 
   const goodResult = await nativeEnumSchema.safeParseAsync(goodData);
   expect(goodResult.success).toBe(true);