--- a/src/__tests__/bigint.test.ts
+++ b/src/__tests__/bigint.test.ts
@@ -1,0 +1,29 @@
+// @ts-ignore TS6133
+import { expect, test } from "@jest/globals";
+
+import { util } from "../helpers/util";
+import * as z from "../index";
+
+test("type guard", () => {
+  const stringToNumber = z.string().transform((arg) => BigInt(Number.MAX_SAFE_INTEGER**(arg.length > 1 ? arg.length : 2)));
+
+  const s1 = z.object({
+    stringToNumber
+  });
+  type t1 = z.input<typeof s1>;
+
+  const data = { stringToNumber: "asdf" };
+  const parsed = s1.safeParse(data);
+  if (parsed.success) {
+    util.assertEqual<typeof data, t1>(true);
+  }
+});
+
+test("test this binding", () => {
+  const callback = (predicate: (val: string) => boolean) => {
+    return predicate("hello");
+  };
+
+  expect(callback((value) => z.string().safeParse(value).success)).toBe(true); // true
+  expect(callback((value) => z.string().safeParse(value).success)).toBe(true); // true
+});