### Type tightening of certain core definitions, namely in `src/helpers/utils.ts`  as follows:


##### arrayToEnum

 - Before

```ts
  export const arrayToEnum = <T extends string, U extends [T, ...T[]]>(
    items: U
  ): { [k in U[number]]: k } => {
    const obj: any = {};
    for (const item of items) {
      obj[item] = item;
    }
    return obj as any;
  };
```

- After

```ts
  export const arrayToEnum = <T extends string, U extends [T, ...T[]]>(
    items: U
  ): { [k in U[number]]: k } => {
    const obj: { [k in U[number]]: k } = Object.prototype.constructor();
    for (const item of items) {
      obj[item] = item;
    }
    return obj as { [k in U[number]]: k };
  };
```


##### getValidEnumValues

- Before

```ts
  export const getValidEnumValues = (obj: any) => {
    const validKeys = objectKeys(obj).filter(
      (k: any) => typeof obj[obj[k]] !== "number"
    );
    const filtered: any = {};
    for (const k of validKeys) {
      filtered[k] = obj[k];
    }
    return objectValues(filtered);
  };
```

- After

```ts
  export const getValidEnumValues = (
    obj: Record<string, string | number>): (string | number)[] => {
    const validKeys: string[] = objectKeys(obj).filter(
      (k) => typeof obj[obj[k]] !== "number"
    );
    const filtered: Record<string, string | number> = {}
    for (const k of validKeys) {
      filtered[k] = obj[k];
    }
    return objectValues(filtered);
  };
```


##### objectValues

- Before

```ts
  export const objectValues = (obj: any) => {
    return objectKeys(obj).map(function (e) {
      return obj[e];
    });
  };
```

- After

```ts
  export const objectValues = (
    (obj: Record<string, string | number>) => {
      return objectKeys(obj).map<string | number>((e, i) => {
        return ({ [i]: obj[e] }[i++]);
      })
    });
```

##### objectKeys

- Before

```ts
  export const objectKeys: ObjectConstructor["keys"] =
    typeof Object.keys === "function" // eslint-disable-line ban/ban
      ? (obj: any) => Object.keys(obj) // eslint-disable-line ban/ban
      : (object: any) => {
          const keys = [];
          for (const key in object) {
            if (Object.prototype.hasOwnProperty.call(object, key)) {
              keys.push(key);
            }
          }
          return keys;
        };
```

- After

```ts
  export const objectKeys: {
    (...args: [o: Record<string, unknown>]): string[];
  } = Array.of<string>()
      ? (obj: Record<string, unknown>) =>
        Object.entries(obj)
          .map(([x, _vals]) => x)
      : (object: Record<string, unknown>) => {
        const keys = Array.of<string>();
        for (const key in object) {
          if (
            Object.prototype.hasOwnProperty.call<
              Record<string, unknown>,
              [string],
              boolean
            >(object, key)
          ) {
            keys.push(key);
          }
        }
        return keys;
      };
```


##### isInteger (performance tweak)

- Before

```ts
  export const isInteger: NumberConstructor["isInteger"] =
    typeof Number.isInteger === "function"
      ? (val) => Number.isInteger(val) // eslint-disable-line ban/ban
      : (val) =>
          typeof val === "number" && isFinite(val) && Math.floor(val) === val;
```

- After

```ts
  export const isInteger: NumberConstructor["isInteger"] =
    typeof Number.isInteger === "function"
      ? (val) => Number.isInteger(val) // eslint-disable-line ban/ban
      : (val) =>
        /// ~~(val: number) === Math.floor(val: number); ~~(val: number) is more performant 
        typeof val === "number" && isFinite(val) && ~~(val) === val;
```