@DopamineDriven 2022-08-28
Reduced testing time substantially by changing the `data` prop in the `getParsedType` function from `any` to `undefined` and handling the `ZodParsedType.promise` conditional in a more strongly typed manner as follows:

##### getParsedType

- Before

 ```ts
 export const getParsedType = (data: any): ZodParsedType => {
  const t = typeof data;

  switch (t) {
    case "undefined":
      return ZodParsedType.undefined;

    case "string":
      return ZodParsedType.string;

    case "number":
      return isNaN(data) ? ZodParsedType.nan : ZodParsedType.number;

    case "boolean":
      return ZodParsedType.boolean;

    case "function":
      return ZodParsedType.function;

    case "bigint":
      return ZodParsedType.bigint;

    case "object":
      if (Array.isArray(data)) {
        return ZodParsedType.array;
      }
      if (data === null) {
        return ZodParsedType.null;
      }
      if (
        data.then &&
        typeof data.then === "function" &&
        data.catch &&
        typeof data.catch === "function"
      ) {
        return ZodParsedType.promise;
      }
      if (typeof Map !== "undefined" && data instanceof Map) {
        return ZodParsedType.map;
      }
      if (typeof Set !== "undefined" && data instanceof Set) {
        return ZodParsedType.set;
      }
      if (typeof Date !== "undefined" && data instanceof Date) {
        return ZodParsedType.date;
      }
      return ZodParsedType.object;

    default:
      return ZodParsedType.unknown;
  }
};
 ```
 
 - After 
 
 ```ts
 
export const getParsedType = (data: unknown): ZodParsedType => {
  const t = typeof data;

  switch (t) {
    case "undefined":
      return ZodParsedType.undefined;

    case "string":
      return ZodParsedType.string;

    case "number":
      return typeof data === "number" && isNaN(data) ? ZodParsedType.nan : ZodParsedType.number;

    case "boolean":
      return ZodParsedType.boolean;

    case "function":
      return ZodParsedType.function;

    case "bigint":
      return ZodParsedType.bigint;

    case "object":
      if (Array.isArray(data)) {
        return ZodParsedType.array;
      }
      if (data === null) {
        return ZodParsedType.null;
      }
      if (
        (data as ReturnType<typeof Promise["any"]>).then &&
        typeof (data as ReturnType<typeof Promise["any"]>).then ===
          "function" &&
        (data as ReturnType<typeof Promise["any"]>).catch &&
        typeof (data as ReturnType<typeof Promise["any"]>).catch === "function"
      ) {
        return ZodParsedType.promise;
      }
      if (typeof Map !== "undefined" && data instanceof Map) {
        return ZodParsedType.map;
      }
      if (typeof Set !== "undefined" && data instanceof Set) {
        return ZodParsedType.set;
      }
      if (typeof Date !== "undefined" && data instanceof Date) {
        return ZodParsedType.date;
      }
      return ZodParsedType.object;

    default:
      return ZodParsedType.unknown;
  }
};
 ```
 
 Only `number` and `promise` required modification after changing the type of `data` from `any` to `unknown` -- using unknown instead over any in the vast majority of cases is preferential for performant type-checking
 
 - `ZodParsedType.number`
 ```ts
     case "number":
      return typeof data === "number" && isNaN(data) ? ZodParsedType.nan : ZodParsedType.number;
 ```
 
 - `ZodParsedType.promise`
 ```ts
      if (
       (data as ReturnType<typeof Promise["any"]>).then &&
       typeof (data as ReturnType<typeof Promise["any"]>).then ===
         "function" &&
       (data as ReturnType<typeof Promise["any"]>).catch &&
       typeof (data as ReturnType<typeof Promise["any"]>).catch === "function"
     ) {
       return ZodParsedType.promise;
     }
 ```
 
 ---
 
 The other area that a small update was made that contributed to an increase in testing speed/performance was in the `src/__tests__/promise.test.ts` file -- the change is shown in the snippet below. Essentially, the `this` keyword needed to have `then` or `catch` chained for each case within the `fakePromise` object
 
 ```ts
 // this reduced testing time substantially (adding .then and .catch to this returned in fakePromise)
test("promise parsing success 2", () => {
  const fakePromise = {
    then() {
      return this.then;
    },
    catch() {
      return this.catch;
    },
  };
  promSchema.parse(fakePromise);
});
 ```
 
 --- 
 
 
 Included below are the test metrics from running `yarn test` -- 3.5 seconds faster than expected after making the slight change to the testing file outlined above 
 
 ```bash
 dopaminedriven@LAPTOP-2IH011V4:~/open/zod$ yarn test
yarn run v1.22.19
$ jest --coverage
 PASS  src/__tests__/promise.test.ts (6.172 s)
 PASS  src/__tests__/discriminatedUnions.test.ts (6.215 s)
 PASS  src/__tests__/error.test.ts (6.559 s)
 PASS  src/__tests__/function.test.ts (6.531 s)
 PASS  src/__tests__/string.test.ts (6.625 s)
 PASS  src/__tests__/nativeEnum.test.ts (6.588 s)
 PASS  src/__tests__/all-errors.test.ts (6.605 s)
 PASS  src/__tests__/transformer.test.ts (6.828 s)
 PASS  src/__tests__/void.test.ts
 PASS  src/__tests__/refine.test.ts (6.856 s)
 PASS  src/__tests__/primitive.test.ts (7.022 s)
 PASS  src/__tests__/object.test.ts
 PASS  src/__tests__/date.test.ts
 PASS  src/__tests__/validations.test.ts
 PASS  src/__tests__/record.test.ts
 PASS  src/__tests__/async-refinements.test.ts
 PASS  src/__tests__/safeparse.test.ts
 PASS  src/__tests__/enum.test.ts
 PASS  src/__tests__/tuple.test.ts
 PASS  src/__tests__/object-augmentation.test.ts
 PASS  src/__tests__/number.test.ts
 PASS  src/__tests__/pickomit.test.ts
 PASS  src/__tests__/instanceof.test.ts
 PASS  src/__tests__/intersection.test.ts
 PASS  src/__tests__/array.test.ts
 PASS  src/__tests__/parser.test.ts
 PASS  src/__tests__/anyunknown.test.ts
 PASS  src/__tests__/firstparty.test.ts
 PASS  src/__tests__/partials.test.ts
 PASS  src/__tests__/set.test.ts
 PASS  src/__tests__/nan.test.ts
 PASS  src/__tests__/complex.test.ts
 PASS  src/__tests__/unions.test.ts
 PASS  src/__tests__/default.test.ts
 PASS  src/__tests__/object-in-es5-env.test.ts
 PASS  src/__tests__/bigint.test.ts
 PASS  src/__tests__/map.test.ts
 PASS  src/__tests__/nullable.test.ts
 PASS  src/__tests__/masking.test.ts
 PASS  src/__tests__/async-parsing.test.ts
 PASS  src/__tests__/parseUtil.test.ts
 PASS  src/__tests__/deepmasking.test.ts
 PASS  src/__tests__/branded.test.ts
 PASS  src/__tests__/base.test.ts
 PASS  src/__tests__/optional.test.ts
 PASS  src/__tests__/mocker.test.ts
 PASS  src/__tests__/description.test.ts
 PASS  src/__tests__/recursive.test.ts
 PASS  src/__tests__/languageServerFeatures.test.ts (9.578 s)
-----------------------------------|---------|----------|---------|---------|------------------------------------------------------------------------------------------------------------------------------
File                               | % Stmts | % Branch | % Funcs | % Lines | Uncovered Line #s                                                                                                            
-----------------------------------|---------|----------|---------|---------|------------------------------------------------------------------------------------------------------------------------------
All files                          |   95.48 |    89.23 |   93.73 |   95.78 |                                                                                                                              
 src                               |   95.95 |    89.52 |   94.57 |   96.14 |                                                                                                                              
  ZodError.ts                      |   91.53 |    84.14 |   83.33 |    91.4 | 154-155,177,188,206,208,210,251,362,418-419                                                                                  
  external.ts                      |     100 |      100 |      50 |     100 |                                                                                                                              
  index.ts                         |     100 |      100 |     100 |     100 |                                                                                                                              
  types.ts                         |   96.42 |    90.75 |   95.63 |   96.66 | ...4,440-443,604,616,835,1108,1716,1774,1901,2032-2033,2309,2327,2480-2487,2708,2727,3030-3031,3264-3270,3527,3535,3768,3785 
 src/__tests__                     |   97.22 |      100 |   94.44 |   97.22 |                                                                                                                              
  Mocker.ts                        |     100 |      100 |     100 |     100 |                                                                                                                              
  crazySchema.ts                   |      80 |      100 |      50 |      80 | 41                                                                                                                           
  languageServerFeatures.source.ts |     100 |      100 |     100 |     100 |                                                                                                                              
 src/helpers                       |    90.9 |     87.5 |   86.84 |    91.8 |                                                                                                                              
  errorUtil.ts                     |     100 |    91.66 |     100 |     100 | 6                                                                                                                            
  parseUtil.ts                     |   97.26 |    94.44 |   94.11 |     100 | 135                                                                                                                          
  util.ts                          |   83.07 |    83.33 |   77.77 |   83.87 | 11,55-67,74-77,89                                                                                                            
-----------------------------------|---------|----------|---------|---------|------------------------------------------------------------------------------------------------------------------------------

Test Suites: 49 passed, 49 total
Tests:       391 passed, 391 total
Snapshots:   0 total
Time:        10.52 s, estimated 14 s
Ran all test suites.
Done in 11.72s.
 ```